Move specific operators closer to the application
