+++ applications/endoscopy_tool_tracking/gxf_extensions/lstm_tensor_rt_inference/metadata.json
