+++ applications/endoscopy_tool_tracking/operators/lstm_tensor_rt_inference/metadata.json
