+++ applications/endoscopy_tool_tracking/operators/lstm_tensor_rt_inference/python/CMakeLists.txt
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2022-2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,9 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib/gxf_extensions)
-set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib/gxf_extensions)
+include(pybind11_add_holohub_module)
+pybind11_add_holohub_module(
+    CPP_CMAKE_TARGET lstm_tensor_rt_inference
+    CLASS_NAME "LSTMTensorRTInferenceOp"
+    SOURCES lstm_tensor_rt_inference.cpp
+)
 
-add_holohub_extension(deltacast_videomaster)
-add_holohub_extension(emergent_source)
-add_holohub_extension(yuan_qcap)
+target_include_directories(lstm_tensor_rt_inference_python PRIVATE ${Holohub_SOURCE_DIR}/operators)
