+++ applications/endoscopy_tool_tracking/operators/tool_tracking_postprocessor/metadata.json
