+++ applications/endoscopy_tool_tracking/operators/tool_tracking_postprocessor/python/tool_tracking_postprocessor.cpp
@@ -25,7 +25,7 @@
 #include <memory>
 #include <string>
 
-#include "../../operator_util.hpp"
+#include "operator_util.hpp"
 #include <holoscan/core/fragment.hpp>
 #include <holoscan/core/operator.hpp>
 #include <holoscan/core/operator_spec.hpp>
