+++ applications/multiai_ultrasound/CMakeLists.txt
@@ -35,6 +35,7 @@ if(HOLOHUB_DOWNLOAD_DATASETS)
   )
 endif()
 
+add_subdirectory(operators)
 add_subdirectory(cpp)
 if(HOLOHUB_BUILD_PYTHON) # since using python wrapping
   add_subdirectory(python)
