+++ applications/multiai_ultrasound/operators/visualizer_icardio/metadata.json
