+++ operators/CMakeLists.txt
@@ -13,15 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-if(HOLOHUB_BUILD_PYTHON)
-  # Define and create a directory to store python operators under the holohub python module
-  if(NOT CMAKE_INSTALL_LIBDIR)
-    set(CMAKE_INSTALL_LIBDIR lib)
-  endif()
-  set(HOLOHUB_PYTHON_MODULE_OUT_DIR ${CMAKE_BINARY_DIR}/python/${CMAKE_INSTALL_LIBDIR}/holohub)
-  file(MAKE_DIRECTORY ${HOLOHUB_PYTHON_MODULE_OUT_DIR})
-endif()
-
 # Add operators (in alphabetical order)
 add_holohub_operator(advanced_network)
 add_holohub_operator(basic_network)
